Fix handling of IPv4 IPs
It seems that net.ParseIP returns a 16 byte array regardless of the 
input IP size.  This isn't true of all code paths, thus we need to be 
careful about not assuming that we'll get a 16 byte array as input.  Now 
we consistently convert ipv4 addresses to 4 byte arrays before operating 
on them.